Refactor constants and collection as modules, and unify sub-modules into cli, as suggested (- WIP #2 -)
