--- piplicenses/cli/custom_help_formatter.py
@@ -1,85 +1,0 @@
-# pip-licenses-cli
-#
-# MIT License
-#
-# Copyright (c) 2018 raimon
-# Copyright (c) 2025 stefan6419846
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-from __future__ import annotations
-
-import argparse
-from enum import Enum
-from typing import TYPE_CHECKING, cast
-
-if TYPE_CHECKING:  # pragma: no cover
-    from typing import Optional
-
-
-def enum_key_to_value(enum_key: Enum) -> str:
-    return enum_key.name.replace("_", "-").lower()
-
-
-class CustomHelpFormatter(argparse.HelpFormatter):  # pragma: no cover
-
-    def __init__(
-        self,
-        prog: str,
-        indent_increment: int = 2,
-        max_help_position: int = 24,
-        width: Optional[int] = None,
-    ) -> None:
-        # Force max_help_position to 30 regardless of the passed-in value
-        max_help_position = 30
-        super().__init__(
-            prog,
-            indent_increment=indent_increment,
-            max_help_position=max_help_position,
-            width=width,
-        )
-
-    def _format_action(self, action: argparse.Action) -> str:
-        flag_indent_argument: bool = False
-        # Expand help to detect markers before default formatting
-        text: str = self._expand_help(action)
-        separator_pos = text[:3].find("|")
-        # If the marker 'I' appears before '|', enable indentation
-        if separator_pos != -1 and "I" in text[:separator_pos]:
-            self._indent()
-            flag_indent_argument = True
-        help_str = super()._format_action(action)
-        if flag_indent_argument:
-            self._dedent()
-        return help_str
-
-    def _expand_help(self, action: argparse.Action) -> str:
-        if isinstance(action.default, Enum):
-            default_value = enum_key_to_value(action.default)
-            return cast(str, self._get_help_string(action)) % {"default": default_value}
-        return super()._expand_help(action)
-
-    def _split_lines(self, text: str, width: int) -> list[str]:
-        separator_pos = text[:3].find("|")
-        if separator_pos != -1:
-            flag_splitlines: bool = "R" in text[:separator_pos]
-            text = text[separator_pos + 1:]  # fmt: skip
-            if flag_splitlines:
-                return text.splitlines()
-        return super()._split_lines(text, width)
