--- piplicenses/cli/select_action.py
@@ -1,72 +1,0 @@
-# pip-licenses-cli
-#
-# MIT License
-#
-# Copyright (c) 2018 raimon
-# Copyright (c) 2025 stefan6419846
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-from __future__ import annotations
-
-from argparse import (
-    Action,
-    ArgumentParser,
-    Namespace,
-)
-from typing import TYPE_CHECKING, Type
-
-from piplicenses_lib import (
-    FromArg,
-    NoValueEnum,
-)
-
-if TYPE_CHECKING:  # pragma: no cover
-    from typing import Optional
-
-from .format_arg import FormatArg
-from .order_arg import OrderArg
-
-
-def value_to_enum_key(value: str) -> str:
-    return value.replace("-", "_").upper()
-
-
-def get_value_from_enum(enum_cls: Type[NoValueEnum], value: str) -> NoValueEnum:
-    return getattr(enum_cls, value_to_enum_key(value))
-
-
-MAP_DEST_TO_ENUM = {
-    "from_": FromArg,
-    "order": OrderArg,
-    "format_": FormatArg,
-}
-
-
-class SelectAction(Action):
-
-    def __call__(  # type: ignore[override]
-        self,
-        parser: ArgumentParser,
-        namespace: Namespace,
-        values: str,
-        option_string: Optional[str] = None,
-    ) -> None:
-        enum_cls = MAP_DEST_TO_ENUM[self.dest]
-        setattr(namespace, self.dest, get_value_from_enum(enum_cls, values))
