+++ piplicenses/output/__init__.py
@@ -25,35 +25,29 @@
 
 from __future__ import annotations
 
-from .colors import output_colored  # noqa: F401
+# ANSI Color Codes
+from ..constants import (
+    SUMMARY_FIELD_NAMES,
+    DEFAULT_OUTPUT_FIELDS,
+    SUMMARY_OUTPUT_FIELDS,
+    FIELDS_TO_METADATA_KEYS,
+)
 
 import sys
 from collections import Counter
-from dataclasses import asdict
 from typing import TYPE_CHECKING, cast
 
 from piplicenses_lib import (
     LICENSE_UNKNOWN,
     FromArg,
 )
-from piplicenses_lib import get_packages as _get_packages
-from piplicenses_lib import (
-    normalize_package_name,
-)
 from prettytable import HRuleStyle, PrettyTable
 
 from ..cli import get_sortby
-from ..cli.custom_namespace import CustomNamespace
-from ..cli.format_arg import FormatArg
+from ..cli import CustomNamespace
+from ..cli import FormatArg
 from ..collection import (
-    case_insensitive_partial_match_set_diff,
-    case_insensitive_partial_match_set_intersect,
-    case_insensitive_set_diff,
-    case_insensitive_set_intersect,
-)
-from ..spdx import (
-    SYSTEM_PACKAGES,
-    _parse_spdx,
+    get_packages,
 )
 
 # Triggers F401 -- Unused -- kept for historical note
@@ -65,9 +59,7 @@ else:  # pragma: no cover
 
 
 if TYPE_CHECKING:  # pragma: no cover
-    from typing import Iterator, Sequence
-
-    from piplicenses_lib import PackageInfo
+    from typing import Sequence
 
 from .csv import CSVPrettyTable
 from .json import (
@@ -77,51 +69,6 @@ from .json import (
 from .plain import PlainVerticalTable
 
 open = open  # allow monkey patching
-
-
-FIELD_NAMES = (
-    "Name",
-    "Version",
-    "License",
-    "LicenseFile",
-    "LicenseText",
-    "NoticeFile",
-    "NoticeText",
-    "Author",
-    "Maintainer",
-    "Description",
-    "URL",
-)
-
-
-SUMMARY_FIELD_NAMES = (
-    "Count",
-    "License",
-)
-
-
-DEFAULT_OUTPUT_FIELDS = (
-    "Name",
-    "Version",
-)
-
-
-SUMMARY_OUTPUT_FIELDS = (
-    "Count",
-    "License",
-)
-
-# Mapping of FIELD_NAMES to METADATA_KEYS where they differ by more than case
-FIELDS_TO_METADATA_KEYS = {
-    "URL": "homepage",
-    "License-Metadata": "license",
-    "License-Classifier": "license_classifier",
-    "LicenseFile": "license_files",
-    "LicenseText": "license_texts",
-    "NoticeFile": "notice_files",
-    "NoticeText": "notice_texts",
-    "Description": "summary",
-}
 
 
 def factory_styled_table_with_args(
@@ -167,105 +114,6 @@ def create_summary_table(args: CustomNamespace) -> PrettyTable:
     for license, count in counts.items():
         table.add_row([count, license])
     return table
-
-
-def get_packages(
-    args: CustomNamespace,
-) -> Iterator[PackageInfo]:
-    ignore_pkgs_as_normalize = [normalize_package_name(pkg) for pkg in args.ignore_packages]
-    pkgs_as_normalize = [normalize_package_name(pkg) for pkg in args.packages]
-
-    fail_on_licenses = set()
-    if args.fail_on:
-        fail_on_licenses = set(map(str.strip, args.fail_on.split(";")))
-
-    allow_only_licenses = set()
-    if args.allow_only:
-        allow_only_licenses = set(map(str.strip, args.allow_only.split(";")))
-
-    include_files = args.with_license_file or args.with_notice_file
-    failures = []
-    for pkg_info in _get_packages(
-        from_source=args.from_,
-        python_path=args.python,
-        include_files=include_files,
-        normalize_names=False,
-    ):
-        pkg_name = normalize_package_name(pkg_info.name)
-        pkg_name_and_version = pkg_name + ":" + pkg_info.version
-
-        if pkg_name.lower() in ignore_pkgs_as_normalize or pkg_name_and_version.lower() in ignore_pkgs_as_normalize:
-            continue
-
-        if pkgs_as_normalize and pkg_name.lower() not in pkgs_as_normalize:
-            continue
-
-        if not args.with_system and pkg_name in SYSTEM_PACKAGES:
-            continue
-
-        if args.filter_strings:
-
-            def filter_string(item: str) -> str:
-                return item.encode(args.filter_code_page, errors="ignore").decode(args.filter_code_page)
-
-            for key, value in asdict(pkg_info).items():
-                if key == "distribution":
-                    continue
-
-                def _handle(_value):
-                    if isinstance(_value, list):
-                        return list(map(_handle, _value))
-                    if isinstance(_value, set):
-                        return set(map(_handle, _value))
-                    if isinstance(_value, tuple):
-                        return tuple(map(_handle, _value))
-                    return filter_string(cast(str, _value))
-
-                setattr(pkg_info, key, _handle(value))
-
-        parsed_license_names: set[str] = set()
-        for license_expr in pkg_info.license_names:
-            parsed_license_names |= _parse_spdx(license_expr)
-
-        fail_this_pkg = False
-        fail_message = ""
-        if fail_on_licenses:
-            if not args.partial_match:
-                failed_licenses = case_insensitive_set_intersect(parsed_license_names, fail_on_licenses)
-            else:
-                failed_licenses = case_insensitive_partial_match_set_intersect(parsed_license_names, fail_on_licenses)
-            if failed_licenses:
-                fail_this_pkg = True
-                fail_message = "fail-on license {} was found for package {}:{}\n".format(
-                    "; ".join(sorted(failed_licenses)),
-                    pkg_info.name,
-                    pkg_info.version,
-                )
-        if allow_only_licenses:
-            if not args.partial_match:
-                uncommon_licenses = case_insensitive_set_diff(parsed_license_names, allow_only_licenses)
-            else:
-                uncommon_licenses = case_insensitive_partial_match_set_diff(parsed_license_names, allow_only_licenses)
-            if len(uncommon_licenses) == len(parsed_license_names):
-                fail_this_pkg = True
-                fail_message = "license {} not in allow-only licenses was found for package {}:{}\n".format(
-                    "; ".join(sorted(uncommon_licenses)),
-                    pkg_info.name,
-                    pkg_info.version,
-                )
-        if fail_this_pkg:
-            if args.collect_all_failures:
-                failures.append(fail_message)
-            else:
-                sys.stderr.write(fail_message)
-                sys.exit(1)
-
-        yield pkg_info
-
-    if args.collect_all_failures and failures:
-        for msg in failures:
-            sys.stderr.write(msg)
-        sys.exit(1)
 
 
 def create_licenses_table(
